feat(ml): add new ml model listing endpoint to platform client
