+++ src/resources/MachineLearning/ModelListing/ModelListing.ts
@@ -1,14 +1,11 @@
 import API from '../../../APICore.js';
 import Resource from '../../Resource.js';
-import {MLListingModel} from './ModelListingConfiguration.js';
+import {MLListingModel} from './ModelListingInterfaces.js';
 
 export default class ModelListing extends Resource {
     static baseUrl = `/rest/organizations/${API.orgPlaceholder}/machinelearning/configuration/modellisting`;
 
     list(engineIds?: string[]) {
-        const requestUrl = engineIds?.length
-            ? `${ModelListing.baseUrl}?${engineIds.map((id) => `engineIds=${encodeURI(id)}`).join('&')}`
-            : ModelListing.baseUrl;
-        return this.api.get<MLListingModel[]>(requestUrl);
+        return this.api.get<MLListingModel[]>(this.buildPath(ModelListing.baseUrl, {engineIds}));
     }
 }
