--- src/resources/MachineLearning/ModelListing/ModelListingConfiguration.ts
@@ -1,24 +1,0 @@
-export interface MLListingModel {
-    modelDisplayName: string;
-    modelId: string;
-    engineId: string;
-    estimatedPreviousModelUpdateTime: number;
-    nextModelUpdateTime: number;
-    platformVersion: 1 | 2;
-    modelSizeStatistic: number;
-    readyForAssociation: boolean;
-    modelStatusInfo: MLModelStatusInfo;
-    modelAssociations: MLModelAssociation[];
-}
-
-export interface MLModelStatusInfo {
-    modelStatus: string;
-    daysUntilArchival?: number;
-}
-
-export interface MLModelAssociation {
-    parentId: string;
-    id: string;
-    name: string;
-    associationType: string;
-}
